requirements.txt
streamlit>=1.28.0
polars>=1.0.0
plotly>=5.18.0
pyserial>=3.5
